Set up travis build with gradle and jar
