+++ build.gradle
@@ -31,5 +31,11 @@ test {
     useJUnitPlatform()
 }
 
+jar {
+    manifest {
+        attributes 'Main-Class': 'CS_348_SP_2019.CIExercise.CIExercise'
+    }
+}
+
 // Define the main class for the application
-mainClassName = 'CS.ci.exercise.CIExercise'
+mainClassName = 'CS_348_SP_2019.CIExercise.CIExercise'
