+++ src/main/java/CS_348_SP_2019/CIExercise/EndsWith3.java
@@ -1,19 +1,18 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
-public class EndsWith9 {
+public class EndsWith3 {
 	
 	String names;
 
-	public EndsWith9()  {
+	public EndsWith3()  {
 		names = "";
 		// add your name here in the form of a line that looks like:
 		// names+= "John Smith\n";
 
 	}
-
+	
 	@Override
 	public String toString() {
-		return "EndsWith9 [names=" + names + "]";
+		return "EndsWith3 [names=" + names + "]";
 	}
-
 }
