+++ src/main/java/CS_348_SP_2019/CIExercise/EndsWith5.java
@@ -1,10 +1,10 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
-public class EndsWith4 {
-	
+public class EndsWith5 {
+
 	String names;
 
-	public EndsWith4()  {
+	public EndsWith5()  {
 		names = "";
 		// add your name here in the form of a line that looks like:
 		// names+= "John Smith\n";
@@ -13,7 +13,7 @@ public class EndsWith4 {
 
 	@Override
 	public String toString() {
-		return "EndsWith4 [names=" + names + "]";
+		return "EndsWith5 [names=" + names + "]";
 	}
 
 }
