+++ src/main/java/CS_348_SP_2019/CIExercise/EndsWith8.java
@@ -1,4 +1,4 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 public class EndsWith8 {
