+++ src/main/java/CS_348_SP_2019/CIExercise/EndsWith9.java
@@ -1,19 +1,19 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
-public class EndsWith5 {
-
+public class EndsWith9 {
+	
 	String names;
 
-	public EndsWith5()  {
+	public EndsWith9()  {
 		names = "";
 		// add your name here in the form of a line that looks like:
 		// names+= "John Smith\n";
-		
+
 	}
 
 	@Override
 	public String toString() {
-		return "EndsWith5 [names=" + names + "]";
+		return "EndsWith9 [names=" + names + "]";
 	}
 
 }
