+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith0Test.java
@@ -1,15 +1,18 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith7Test {
+import CS_348_SP_2019.CIExercise.EndsWith0;
+
+public class EndsWith0Test {
 
 	@Test
 	public void testToString() {
-		EndsWith7 seven = new EndsWith7();
+		EndsWith0 zero = new EndsWith0();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith7 [names=]", seven.toString());
+		assertEquals("EndsWith0 [names=]", zero.toString());
+
 	}
 
 }
