+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith1Test.java
@@ -1,16 +1,18 @@
 
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith4Test {
+import CS_348_SP_2019.CIExercise.EndsWith1;
+
+public class EndsWith1Test {
 
 	@Test
 	public void testToString() {
-		EndsWith4 four = new EndsWith4();
+		EndsWith1 one = new EndsWith1();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith4 [names=]", four.toString());
+		assertEquals("EndsWith1 [names=]", one.toString());
 	}
 
 }
