+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith2Test.java
@@ -1,16 +1,18 @@
 
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith1Test {
+import CS_348_SP_2019.CIExercise.EndsWith2;
+
+public class EndsWith2Test {
 
 	@Test
 	public void testToString() {
-		EndsWith1 one = new EndsWith1();
+		EndsWith2 two = new EndsWith2();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith1 [names=]", one.toString());
+		assertEquals("EndsWith2 [names=Karl Wurst\n]", two.toString());
 	}
 
 }
