+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith3Test.java
@@ -1,15 +1,18 @@
-package CS_348_SP_2018.CIExercise;
+
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith0Test {
+import CS_348_SP_2019.CIExercise.EndsWith3;
+
+public class EndsWith3Test {
 
 	@Test
 	public void testToString() {
-		EndsWith0 zero = new EndsWith0();
+		EndsWith3 three = new EndsWith3();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith0 [names=]", zero.toString());
+		assertEquals("EndsWith3 [names=]", three.toString());
 
 	}
