+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith4Test.java
@@ -1,18 +1,18 @@
 
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith8Test {
+import CS_348_SP_2019.CIExercise.EndsWith4;
+
+public class EndsWith4Test {
 
 	@Test
 	public void testToString() {
-		EndsWith8 eight = new EndsWith8();
+		EndsWith4 four = new EndsWith4();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith8 [names=]", eight.toString());
-
+		assertEquals("EndsWith4 [names=]", four.toString());
 	}
 
 }
-
