+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith5Test.java
@@ -1,15 +1,17 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith9Test {
+import CS_348_SP_2019.CIExercise.EndsWith5;
+
+public class EndsWith5Test {
 
 	@Test
 	public void testToString() {
-		EndsWith9 nine = new EndsWith9();
+		EndsWith5 five = new EndsWith5();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith9 [names=]", nine.toString());
+		assertEquals("EndsWith5 [names=]", five.toString());
 
 	}
