+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith7Test.java
@@ -1,16 +1,17 @@
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith5Test {
+import CS_348_SP_2019.CIExercise.EndsWith7;
+
+public class EndsWith7Test {
 
 	@Test
 	public void testToString() {
-		EndsWith5 five = new EndsWith5();
+		EndsWith7 seven = new EndsWith7();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith5 [names=]", five.toString());
-
+		assertEquals("EndsWith7 [names=]", seven.toString());
 	}
 
 }
