+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith8Test.java
@@ -1,16 +1,20 @@
 
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith2Test {
+import CS_348_SP_2019.CIExercise.EndsWith8;
+
+public class EndsWith8Test {
 
 	@Test
 	public void testToString() {
-		EndsWith2 two = new EndsWith2();
+		EndsWith8 eight = new EndsWith8();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith2 [names=]", two.toString());
+		assertEquals("EndsWith8 [names=]", eight.toString());
+
 	}
 
 }
+
