+++ src/test/java/CS_348_SP_2019/CIExercise/EndsWith9Test.java
@@ -1,16 +1,17 @@
-
-package CS_348_SP_2018.CIExercise;
+package CS_348_SP_2019.CIExercise;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
-public class EndsWith3Test {
+import CS_348_SP_2019.CIExercise.EndsWith9;
+
+public class EndsWith9Test {
 
 	@Test
 	public void testToString() {
-		EndsWith3 three = new EndsWith3();
+		EndsWith9 nine = new EndsWith9();
 		// add your name to the expected string, after the all the other names in the form John Smith\n
-		assertEquals("EndsWith3 [names=]", three.toString());
+		assertEquals("EndsWith9 [names=]", nine.toString());
 
 	}
